+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.56.0] - Unpublished
+## [0.56.0] - 2024-04-02
 
 ### Added
 
 - `walletg` in samples providing CLI using Sui GraphQL RPC
+  - Includes dry run for either TransactionData or TransactionKind
+  - Includes executing signed transactions
 
 ### Fixed